Initial commit
This script is a quick and dirty helper when working with ruby app to get data from one place to
It returns an array of hash from a csv file. The hash
keys are the column header symbolized. It returns one hash per row.
another through a console.
It includes a convinience method #pbcopy to get te output array copied to the
clipboard on osx.